refactor to consolidate select options in form into reusable components
