+++ code/src/ui/interaction/form/field/form-field.tsx
@@ -1,16 +1,13 @@
-import generalStyles from '../form.module.css';
 import styles from './field.module.css';
 
-import { useState } from 'react';
 import { UseFormReturn } from 'react-hook-form';
 
 import { PropertyShape, SEARCH_FORM_TYPE, VALUE_KEY } from 'types/form';
-import LoadingSpinner from 'ui/graphic/loader/spinner';
 import { FORM_STATES } from '../form-utils';
 import FormDateTimePicker from './form-date-time-picker';
 import FormInputField from './form-input';
 import FormInputMinMaxField from './input/form-min-max-input';
-import FormSelector from './input/form-selector';
+import OntologyConceptSelector from './input/ontology-concept-selector';
 
 interface FormFieldProps {
   agentApi?: string;
@@ -31,8 +28,6 @@ interface FormFieldProps {
  */
 export default function FormFieldComponent(props: Readonly<FormFieldProps>) {
   const formType: string = props.form.getValues(FORM_STATES.FORM_TYPE);
-  // For dropdown fetching
-  const [isFetching, setIsFetching] = useState<boolean>(true);
   // Any id field in the search form should be ignored
   if (!(formType == SEARCH_FORM_TYPE && props.field.name[VALUE_KEY] == "id")) {
     if (props.field.datatype && ["string", "integer", "decimal"].includes(props.field.datatype)) {
@@ -81,17 +76,11 @@ export default function FormFieldComponent(props: Readonly<FormFieldProps>) {
     } else if (props.field.in) {
       return (
         <div className={styles["form-field-container"]}>
-          {isFetching && (
-            <div className={generalStyles["loader-container"]}>
-              <LoadingSpinner isSmall={true} />
-            </div>)
-          }
           <div className={styles["form-input-container"]}>
-            <FormSelector
+            <OntologyConceptSelector
               agentApi={props.agentApi}
               field={props.field}
               form={props.form}
-              setIsFetching={setIsFetching}
               styles={{
                 label: [styles["form-input-label"]],
               }}
